# docs/docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: MagiDesk POS API
  description: |
    Complete API documentation for MagiDesk Point of Sale System.
    
    The API is versioned with two versions:
    - **v1**: Legacy endpoints (backward compatible, no RBAC enforcement)
    - **v2**: RBAC-enabled endpoints (requires permissions)
    
    All v2 endpoints require authentication via `X-User-Id` header and enforce role-based access control.
  version: 2.0.0
  contact:
    name: MagiDesk Development Team
    url: https://github.com/zedfauji/pos-app
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT

servers:
  - url: https://magidesk-users-904541739138.northamerica-south1.run.app
    description: UsersApi Production
  - url: https://magidesk-menu-904541739138.northamerica-south1.run.app
    description: MenuApi Production
  - url: https://magidesk-order-904541739138.northamerica-south1.run.app
    description: OrderApi Production
  - url: https://magidesk-payment-904541739138.northamerica-south1.run.app
    description: PaymentApi Production
  - url: https://magidesk-inventory-904541739138.northamerica-south1.run.app
    description: InventoryApi Production
  - url: https://magidesk-settings-904541739138.northamerica-south1.run.app
    description: SettingsApi Production
  - url: https://magidesk-customer-904541739138.northamerica-south1.run.app
    description: CustomerApi Production
  - url: https://magidesk-discount-904541739138.northamerica-south1.run.app
    description: DiscountApi Production
  - url: https://magidesk-tables-904541739138.northamerica-south1.run.app
    description: TablesApi Production

tags:
  - name: Authentication
    description: User authentication and session management
  - name: Users
    description: User management operations
  - name: RBAC
    description: Role-Based Access Control operations
  - name: Menu
    description: Menu item management
  - name: Orders
    description: Order processing and management
  - name: Payments
    description: Payment processing operations
  - name: Inventory
    description: Inventory management
  - name: Settings
    description: System settings management
  - name: Customers
    description: Customer management
  - name: Discounts
    description: Discount and promotion management
  - name: Tables
    description: Table and session management

paths:
  /api/v2/auth/login:
    post:
      tags:
        - Authentication
      summary: User login
      description: Authenticate user and return session information including permissions
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Invalid credentials
        '500':
          description: Server error

  /api/v2/users:
    get:
      tags:
        - Users
      summary: Get all users
      description: Retrieve paginated list of users (requires user:view permission)
      operationId: getUsers
      parameters:
        - name: SearchTerm
          in: query
          schema:
            type: string
        - name: Role
          in: query
          schema:
            type: string
        - name: IsActive
          in: query
          schema:
            type: boolean
        - name: Page
          in: query
          schema:
            type: integer
            default: 1
        - name: PageSize
          in: query
          schema:
            type: integer
            default: 20
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedUsersResponse'
        '403':
          description: Forbidden - missing permission
        '500':
          description: Server error
    post:
      tags:
        - Users
      summary: Create user
      description: Create a new user (requires user:create permission)
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      security:
        - BearerAuth: []
      responses:
        '201':
          description: User created
        '403':
          description: Forbidden - missing permission
        '500':
          description: Server error

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    UserIdHeader:
      type: apiKey
      in: header
      name: X-User-Id

  schemas:
    LoginRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
        password:
          type: string
          format: password

    LoginResponse:
      type: object
      properties:
        userId:
          type: string
          format: uuid
        username:
          type: string
        role:
          type: string
        permissions:
          type: array
          items:
            type: string
        lastLoginAt:
          type: string
          format: date-time

    PagedUsersResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/UserDto'
        totalCount:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer

    UserDto:
      type: object
      properties:
        userId:
          type: string
          format: uuid
        username:
          type: string
        role:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    CreateUserRequest:
      type: object
      required:
        - username
        - password
        - role
      properties:
        username:
          type: string
        password:
          type: string
          format: password
        role:
          type: string
        isActive:
          type: boolean
          default: true